Added BMP388 class. Also introduced some issues (check description)
For some reason, now the code only saves to the board like 50% of the time. Oftentimes you have to unplug the board and plug it back in to be able to flash the code to the board. Not sure why. Furthermore, if both BMP388 and ICM42605 are running, the thread breaks. Again, no idea why